# [java] Make CloneMethodMustImplementCloneable over 500x faster