- This was an excruciatingly slow rule, the slowest by far of all PMD
 - We speed it up by:
    * checking early if we are analyzing a `clone()` method to cut the AST tree
    * making better use of type information available
    * only performing additional checks when we know we can't rely on
       available type information

# Benchmark on PMD sourcecode

### -d ./ -f text -R java-typeresolution -b -r report-oldclone.text -threads 0 -auxclasspath ...

Original code:
```
---------------------------------<<< Rules >>>---------------------------------
Rule name                                       Time (secs)    # of Evaluations

LooseCoupling:                                        0.215               1,643
SignatureDeclareThrowsException:                      0.279               1,643
UnusedImports:                                        0.356               1,643
CloneMethodMustImplementCloneable:                   25.642               1,643
```

This PR:

```
---------------------------------<<< Rules >>>---------------------------------
Rule name                                       Time (secs)    # of Evaluations

CloneMethodMustImplementCloneable:                    0.048               1,643
LooseCoupling:                                        0.309               1,643
SignatureDeclareThrowsException:                      0.342               1,643
UnusedImports:                                        0.518               1,643
```